backend/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
black==22.12.0
boto3==1.26.34
botocore==1.29.34
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cognitojwt==1.4.1
cryptography==38.0.4
dnspython==2.2.1
ecdsa==0.18.0
fastapi==0.88.0
google-auth==2.15.0
h11==0.14.0
httptools==0.5.0
idna==3.3
jmespath==1.0.1
motor==3.1.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
requests==2.28.0
rsa==4.9
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
starkbank-ecdsa==2.2.0
starlette==0.22.0
tomli==2.0.1
twilio==7.16.0
typing_extensions==4.4.0
urllib3==1.26.9
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
